Spline modification info added.
The manipulation methods in the spline class now return information
about which interpolated points were affected by the manipulation. This
information can be used to correctly replace and recalculate the affected section of the mesh with the new interpolated points. The deleteControlPoint method is not optimal currently, it recalculates more points than strictly necessary. The result appears to be correct though.