Spline interface + improvement of mesh generation.
The replacePoints method of the LineExtruder class was reworked to now
also recalculate adjacent points to the replaced section. This is
required to make sure the orientation of the cross sections is correct
after replacing a section.
It also turned out that the replacePoints method is versatile enough that
the getMesh implementation could be replaced by a call to it.
A spline interface was added that the LinearInterpolationSpline and
KochanekBartelsSpline implement that makes them easily interchangeable.