Cross sections of the line are projected for constant diameter.
Before the cross sections were just translated and rotated which meant
that the diameter of the mesh sections was not constant. Now it is
projected using ray casts to a plane at the point of the line. There are
still issues with twisting of the mesh and the normals are not yet
correctly calculated. The automatic normal calculation function of Unity
could also be used here.