Last segment of spline now reaches last control point.
Calculation of the last segment was changed to reach the position of the
last control point. Before it would end one interpolated point before because
the first point of the next segment would be at the position and the
last segment was not treated differently.
Also refactored the method getControlPointsForSegment.