LineSketchObject class added.
Some minor changes to the Spline classes were made to make the number of
control points available. The LineSketchObject was added to allow
interaction with a specific line in the scene. It allows adding and
deleting control points at the end of the line and updates the mesh
collider with the generated mesh of the line. The LineSketchingTool
allows interaction with all LineSketchObjects in the scene but
is specific to the ARCore plugin for Unity so it will not work without
it.