Changes to selection, sketch world, sketch object group.
Sketch objects and groups are now parented to selections so their
position can be easily transformed. Added a reference to the parent
group to sketch objects and groups so they can be deselected and are in
the right place in the hierarchy again. Added a bin object for deleted
objects in the sketch world component. Introduced namespace
SketchObjectManagement.