Introduced ISelectable and SelectableObject abstract class.
They aggregate IHighlightable and IGroupable. SelectableObject is also a
MonoBehaviour. Changed Selection and Group class to use
SelectableObject.

Also changed commands so that when the first control point is deleted
the whole sketch object is deleted. Fixed behaviour of Ribbon. Mesh is
null when there are less than 2 control points. Added test cases for
this.